backend/requirements.txt
# FastAPI + Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# LangChain + LangGraph
langchain==0.3.0
langchain-core==0.3.0
langchain-ollama==0.2.0
langchain-google-genai==2.0.0
langgraph==0.2.28

# Neo4j
neo4j==5.25.0

# GraphRAG components (ファイルモード用、Neo4jモードでは不要)
networkx==3.3
# hnswlib: pip install chroma-hnswlib
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
